+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Development Version:
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
 - Rewrite _unlabeled_dimension_mix (labeled/unlabeled), add tests.
   By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
+- Add default netcdf fillvalues, pad only if necessary, adapt tests.
+  By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_.
 
 Version 1.0.2 (August 2nd, 2022):